+++ packages/react-vapor/src/components/slider/tests/MiddleSlider.spec.tsx
@@ -7,6 +7,7 @@ import {Provider} from 'react-redux';
 
 import {getStoreMock, ReactVaporMockStore} from '../../../utils/tests/TestUtils';
 import {MiddleSlider} from '../MiddleSlider';
+import {computeStep} from '../SliderUtils';
 
 describe('<MiddleSlider/>', () => {
     let store: ReactVaporMockStore;
@@ -15,6 +16,7 @@ describe('<MiddleSlider/>', () => {
         max: 100,
         min: -100,
     };
+    const step = 50;
 
     beforeEach(() => {
         store = getStoreMock({
@@ -38,7 +40,7 @@ describe('<MiddleSlider/>', () => {
                 customTooltip={() => <span>customTooltip</span>}
                 hasTooltip
                 marks={{0: '-2000', 33: '2000', 17: '0', 100: '10,000'}}
-                step={5}
+                step={step}
                 initialValue={20}
                 enabled
             />,
@@ -60,7 +62,7 @@ describe('<MiddleSlider/>', () => {
                 onChange: jasmine.any(Function),
                 hasTooltip: true,
                 marks: {0: '-2000', 33: '2000', 17: '0', 100: '10,000'},
-                step: 5,
+                step,
                 enabled: true,
                 initialValue: 20,
             };
@@ -71,24 +73,28 @@ describe('<MiddleSlider/>', () => {
         it('should render a track it with its marks', () => {
             middleSlider = shallowedMiddleSlider()
                 .dive()
+                .childAt(0)
                 .dive();
             const children: any = middleSlider.prop('children');
             const marks = children[2].props.marks;
             expect(marks).toEqual({0: '-2000', 33: '2000', 17: '0', 100: '10,000'});
         });
 
-        it('should apply the step prop', () => {
+        it('should apply the computed step prop to the range slider', () => {
             middleSlider = shallowedMiddleSlider()
                 .dive()
+                .childAt(0)
                 .dive();
             const children: any = middleSlider.prop('children');
-            const step = children[2].props.step;
-            expect(step).toEqual(5);
+            const computedStep = computeStep(step, middleSliderRequiredProps.min, middleSliderRequiredProps.max);
+            const middleSliderStep = children[2].props.step;
+            expect(middleSliderStep).toEqual(computedStep);
         });
 
         it('should set the handle hasTooltip prop to true when slider hasTooltip his passed', () => {
             middleSlider = shallowedMiddleSlider()
                 .dive()
+                .childAt(0)
                 .dive();
             const children: any = middleSlider.prop('children');
             const firstHandle = children[3][0];
@@ -109,6 +115,7 @@ describe('<MiddleSlider/>', () => {
                 store
             )
                 .dive()
+                .childAt(0)
                 .dive();
             const children: any = middleSlider.prop('children');
             const firstHandle = children[3][0];
@@ -118,6 +125,7 @@ describe('<MiddleSlider/>', () => {
         it('should not pass the disabled props to the handle if the enable prop is set to true', () => {
             middleSlider = shallowedMiddleSlider()
                 .dive()
+                .childAt(0)
                 .dive();
             const children: any = middleSlider.prop('children');
             const firstHandle = children[3][0];
@@ -138,11 +146,48 @@ describe('<MiddleSlider/>', () => {
                 store
             )
                 .dive()
+                .childAt(0)
                 .dive();
 
             const children: any = middleSlider.prop('children');
             const firstHandle = children[3][0];
             expect(firstHandle.props.handleProps.disabled).toBeTruthy();
+        });
+
+        it('should append the computed  value if the appendedValue prop is passed', () => {
+            middleSlider = shallowWithStore(
+                <MiddleSlider
+                    {...middleSliderRequiredProps}
+                    onChange={(value) => value}
+                    initialValue={20}
+                    customTooltip={() => <span>customTooltip</span>}
+                    marks={{0: '-2000', 33: '2000', 17: '0', 100: '10,000'}}
+                    step={5}
+                    enabled={false}
+                    appendValue
+                />,
+                store
+            ).dive();
+
+            expect(middleSlider.childAt(0).prop('className')).toContain('appended-value');
+            expect(middleSlider.childAt(1).prop('className')).not.toContain('hidden');
+        });
+        it('should NOT append the computed  value if the appendedValue prop is passed', () => {
+            middleSlider = shallowWithStore(
+                <MiddleSlider
+                    {...middleSliderRequiredProps}
+                    onChange={(value) => value}
+                    initialValue={20}
+                    customTooltip={() => <span>customTooltip</span>}
+                    marks={{0: '-2000', 33: '2000', 17: '0', 100: '10,000'}}
+                    step={5}
+                    enabled={false}
+                />,
+                store
+            ).dive();
+
+            expect(middleSlider.childAt(0).prop('className')).not.toContain('appended-value');
+            expect(middleSlider.childAt(1).prop('className')).toContain('hidden');
         });
     });